Move android specific (nothing to do with our app) tsest helpers to a package
