+++ audiorecorder/src/test/java/org/odk/collect/audiorecorder/recording/AudioRecorderTest.kt
@@ -5,7 +5,7 @@ import org.hamcrest.Matchers.equalTo
 import org.junit.After
 import org.junit.Test
 import org.odk.collect.audiorecorder.recorder.Output
-import org.odk.collect.testshared.LiveDataTester
+import org.odk.collect.testshared.android.LiveDataTester
 import java.io.File
 import java.io.Serializable
