+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formentry/FieldListUpdateTest.java
@@ -40,7 +40,7 @@ import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.core.IsNot.not;
 import static org.hamcrest.core.StringStartsWith.startsWith;
 import static org.odk.collect.android.support.matchers.CustomMatchers.withIndex;
-import static org.odk.collect.testshared.NestedScrollToAction.nestedScrollTo;
+import static org.odk.collect.testshared.android.NestedScrollToAction.nestedScrollTo;
 
 import android.Manifest;
 import android.app.Activity;
@@ -65,11 +65,11 @@ import org.odk.collect.android.R;
 import org.odk.collect.android.TestSettingsProvider;
 import org.odk.collect.android.preferences.GuidanceHint;
 import org.odk.collect.android.storage.StoragePathProvider;
+import org.odk.collect.android.support.pages.FormEntryPage;
 import org.odk.collect.android.support.rules.FormActivityTestRule;
 import org.odk.collect.android.support.rules.ResetStateRule;
-import org.odk.collect.android.support.pages.FormEntryPage;
 import org.odk.collect.settings.keys.ProjectKeys;
-import org.odk.collect.testshared.RecordedIntentsRule;
+import org.odk.collect.testshared.android.RecordedIntentsRule;
 
 import java.io.File;
 import java.io.FileOutputStream;
