+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/projects/AddNewProjectTest.kt
@@ -8,7 +8,7 @@ import org.odk.collect.android.support.TestDependencies
 import org.odk.collect.android.support.pages.MainMenuPage
 import org.odk.collect.android.support.rules.CollectTestRule
 import org.odk.collect.android.support.rules.TestRuleChain
-import org.odk.collect.testshared.RecordedIntentsRule
+import org.odk.collect.testshared.android.RecordedIntentsRule
 
 class AddNewProjectTest {
