+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/settings/FormManagementSettingsTest.kt
@@ -15,7 +15,7 @@ import org.odk.collect.android.support.TestDependencies
 import org.odk.collect.android.support.pages.MainMenuPage
 import org.odk.collect.android.support.rules.CollectTestRule
 import org.odk.collect.android.support.rules.TestRuleChain
-import org.odk.collect.testshared.RecordedIntentsRule
+import org.odk.collect.testshared.android.RecordedIntentsRule
 
 @RunWith(AndroidJUnit4::class)
 class FormManagementSettingsTest {
