+++ collect_app/src/test/java/org/odk/collect/android/external/FormUriActivityTest.kt
@@ -36,8 +36,8 @@ import org.odk.collect.projects.Project
 import org.odk.collect.projects.ProjectsRepository
 import org.odk.collect.settings.SettingsProvider
 import org.odk.collect.shared.strings.UUIDGenerator
-import org.odk.collect.testshared.ActivityScenarioLauncherRule
-import org.odk.collect.testshared.RecordedIntentsRule
+import org.odk.collect.testshared.android.ActivityScenarioLauncherRule
+import org.odk.collect.testshared.android.RecordedIntentsRule
 
 @RunWith(AndroidJUnit4::class)
 class FormUriActivityTest {
