+++ geo/src/test/java/org/odk/collect/geo/GeoPointMapActivityTest.java
@@ -23,7 +23,7 @@ import org.odk.collect.externalapp.ExternalAppUtils;
 import org.odk.collect.geo.maps.MapFragmentFactory;
 import org.odk.collect.geo.maps.MapPoint;
 import org.odk.collect.geo.support.FakeMapFragment;
-import org.odk.collect.testshared.ActivityScenarioLauncherRule;
+import org.odk.collect.testshared.android.ActivityScenarioLauncherRule;
 import org.robolectric.shadows.ShadowApplication;
 
 @RunWith(AndroidJUnit4.class)
