+++ geo/src/test/java/org/odk/collect/geo/GeoPolyActivityTest.java
@@ -38,7 +38,7 @@ import org.mockito.Mockito;
 import org.odk.collect.geo.maps.MapFragmentFactory;
 import org.odk.collect.geo.support.FakeMapFragment;
 import org.odk.collect.location.tracker.LocationTracker;
-import org.odk.collect.testshared.ActivityScenarioLauncherRule;
+import org.odk.collect.testshared.android.ActivityScenarioLauncherRule;
 import org.robolectric.shadows.ShadowApplication;
 
 @RunWith(AndroidJUnit4.class)
