+++ location/src/test/java/org/odk/collect/location/tracker/LocationTrackerServiceTest.kt
@@ -7,7 +7,7 @@ import org.hamcrest.Matchers.notNullValue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.odk.collect.androidshared.ui.ReturnToAppActivity
-import org.odk.collect.testshared.ServiceScenario
+import org.odk.collect.testshared.android.ServiceScenario
 
 @RunWith(AndroidJUnit4::class)
 class LocationTrackerServiceTest {
